+++ src/Filters/Concerns/HasLikeOperator.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Konekt\AppShell\Filters\Concerns;
+
+trait HasLikeOperator
+{
+    protected string $likeOperator = 'like';
+
+    public function useILikeOperator(): self
+    {
+        $this->likeOperator = 'ilike';
+
+        return $this;
+    }
+}